builder changes

+++ src/progression/fixed.rs
@@ -1,2 +1,2 @@
-pub const MIN_LEVEL: usize = 1;
-pub const MAX_LEVEL: usize = 6;
+pub const MIN_LEVEL: u32 = 1;
+pub const MAX_LEVEL: u32 = 6;
